feature(options) Initialize options on install
Options are not initialized after extnesion installation. The user
is not aware of this and it looks like the extension doesn't work.

To address the issue:
 * Open opetions page after onInstalled event
 * Load default options
 * Save options

The drawback of the new solution is that options are automatically
saved every time the options are opened, not just after install.
However, that is a low price for the simplicity.